21_텔레그램으로 나만의 비서만들기/requirements.txt
python-telegram-bot
selenium
webdriver-manager
